requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
greenlet
alembic
pydantic
pydantic-settings
python-dotenv
google-genai
chromadb
sentence-transformers
datasets
mcp[cli]
slowapi
pytest
pytest-asyncio
